WebSocket url fixes.

+++ main-service/src/main/kotlin/tv/comnata/mainservice/entities/websocket/requests/RoomActionRequest.kt
@@ -3,7 +3,7 @@ package tv.comnata.mainservice.entities.websocket.requests
 import java.time.LocalDateTime
 
 class RoomActionRequest {
-    val seekTime: Int? = null
+    val seekTime: Long? = null
     val actionTime: LocalDateTime? = null
     val type: String? = null
 }
